ingredient reducer added
