+++ src/redux/store.js
@@ -1,9 +1,10 @@
 import { createStore, combineReducers } from 'redux';
-import { dataReducer, recipeReducer } from './reducers/index';
+import { dataReducer, recipeReducer, ingredientReducer } from './reducers/index';
 
 const storeReducer = () => createStore(combineReducers({
   datas: dataReducer,
   recipes: recipeReducer,
+  ingredients: ingredientReducer,
 }));
 
 export default storeReducer;
